split OggStream class into: MusicStream (OpenAL) and Decoder
